+++ features/object_removal/__init__.py
@@ -1,0 +1,1 @@
+﻿"""object removal feature package."""